Refactor into cleaner files to prep for shape and purpose impl
